cpc-agent-server/requirements.txt
claude-agent-sdk
supabase
httpx
python-dotenv
